refactor: recyclerView의 Holder 클래스 파일로 분리
